+++ services/sso/src/openapi/components/errorSchemas.ts
@@ -56,9 +56,17 @@ export const ValidationError = registry.register(
     ),
   })
 )
+
 export const InvalidCredentials = registry.register(
   'InvalidCredentialsError',
   z.object({
     message: z.string().openapi({ example: 'Invalid Credentials' }),
+  })
+)
+
+export const DeletedUser = registry.register(
+  'DeletedUserError',
+  z.object({
+    message: z.string().openapi({ example: 'User already deleted' }),
   })
 )