+++ services/sso/src/schemas/users/userSchema.ts
@@ -27,6 +27,7 @@ export const userSchema = z.object({
   status: userStatusSchema,
   createdAt: z.string().datetime(),
   updatedAt: z.string().datetime(),
+  deletedAt: z.string().datetime(),
   itineraryId: z.string().cuid2(),
 })